Made an initial stab at a gnucash backend module.
NOTE! This is only tested to compile. It has not been tested to actually
work.

Also, the backend does not return a gnucash guid for created
transactions (as it should), and this means that
remove_backend_transaction also doesn't do any thing.

Additional note, the gnucash_file attribute has to be set to make this
useful. The ability to set this and any other arbitrary backend
attribute should be added to the script bo-keep-module-control (hint,
use the setattr function)

--HG--
extra : convert_revision : svn%3A478ed200-3351-0410-a957-86bb2dd1a256/bo-keep/trunk%40274