commited a script that repairs the borked payroll database from bokeep 0.3
A good sanity check for this would be to create a bunch of payrolls,
some of them duplicates, and then to find the YTD values and to look and
see if the add up with the ones in the books. An easy YTD value to check
is the vacation pay, pay it all out and look at the balance.

--HG--
extra : convert_revision : svn%3A478ed200-3351-0410-a957-86bb2dd1a256/bo-keep/trunk%40351