More work towards having bo-keep take care of creating its own config file (if not present), and creating or giving the user the chance to select where the database is kept [the only thing kept in the config file].
bo-keep has support for command line processing now, first option is
overiding the config file.

cmdline interface intended to be replaced with real gui.

The later interface (selction of database) will be made possible under
regular circumstances within bokeep as well.

Beyond these two goals, there will need to be
a gui for managing which books are manged, which backend modules they
each use, and the creation of instances of bokeep plugins that they use.

This is not fully tested, but I'm expected to be in the only developer
in the  next few days, so I'm now going to bother branching, just going
to get this right.

--HG--
extra : convert_revision : svn%3A478ed200-3351-0410-a957-86bb2dd1a256/bo-keep/trunk%40665