* fixed full stack test to be compatible with gnucash 2.2 * fixed bugs in trustor_managment re new account selection feature * fixed bugs in gnucash backend plugins where gnucash_file wasn't specified  right  after being changed in config dialog * enhanced trust plugin to take advantage of date/description attributes supported by gnucash backend
This is a pretty significant commit, this is the first time the full stack
from the trust plugin, through the bo-keep core, to the gnucash backend
plugin has been available, useable, and configurable entirely in the
bo-keep gui. High fives!

These means we're super close to tagging this as 1.0. (even though there are
bugs!)
All that remains is to have an About dialog added, setup.py tested out to
see if we can run this thing outside the farm, and a bo-keep.desktop file
created and distributed to make the app easilly launchable.

--HG--
extra : convert_revision : svn%3A478ed200-3351-0410-a957-86bb2dd1a256/bo-keep/trunk%40713