Added a new method to bokeepdb.py that searches the PYTHONPATH and subdirectories called /bokeep/plugins/ for modules and packages marked as BoKeep plugins.
Autopopulated the combo entry box in the plugin add area of the config
dialog with the names of autodetected plugins, so the user doesn't have
to know front-end plugin names any longer.

Changed README to reflect these changes, and also reorganized and divided
the file into sections so it's more clear.