made gnucash backend tests behave more like how bokeep.book.Book would call a backend plugin in order to better teat out bug #33616
So now the tests that all passsed, all fail, and we know this is the gnucash backend plugin's fault

But also wanted to preserve how these tests were behaving before,
because that behavior was also valid and not failing, so there was no
desire to lose a perfectly reasonable test sequence.

So, did inheritance meyhem, and a long story made short we now have
twice the tests for waaaay less than twice the code.

Only regret for this commit is how I kludged the setUp() method resolution order

--HG--
extra : rebase_source : 0e0495e10896a251f924f971b7a9a171237109fb